+++ setup/install_nvidia_docker.sh
@@ -1,18 +1,19 @@
 #!/bin/bash
 set -e
 
-sudo install -m 0755 -d /etc/apt/keyrings
+if [ ! -f /etc/apt/keyrings/nvidia-container-toolkit-keyring.gpg ]; then
+    curl -fsSLk https://nvidia.github.io/libnvidia-container/gpgkey | \
+        sudo gpg --dearmor -o /etc/apt/keyrings/nvidia-container-toolkit-keyring.gpg
+    sudo chmod a+r /etc/apt/keyrings/nvidia-container-toolkit-keyring.gpg
+fi
 
-curl -fsSLk https://nvidia.github.io/libnvidia-container/gpgkey | \
-    sudo gpg --dearmor -o /etc/apt/keyrings/nvidia-container-toolkit-keyring.gpg
-sudo chmod a+r /etc/apt/keyrings/nvidia-container-toolkit-keyring.gpg
-
-curl -sLk https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list | \
-    sed 's#deb https://#deb [signed-by=/etc/apt/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' | \
-    sudo tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
+if [ ! -f /etc/apt/sources.list.d/nvidia-container-toolkit.list ]; then
+    curl -sLk https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list | \
+        sed 's
+        sudo tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
+fi
 
 sudo apt-get update
 sudo apt-get install -y nvidia-container-toolkit
-
 sudo nvidia-ctk runtime configure --runtime=docker
 sudo systemctl restart docker
